factor out normalizeFile
